requirements.txt
flask
flask-sqlalchemy
flask-security-too
flask-admin